Move `ExceptionMessages` into `exceptions` folder
